+++ src/tesseract_main.cc
@@ -181,9 +181,9 @@ struct Args {
     {
       config.det_orders.resize(num_det_orders);
       std::mt19937_64 rng(det_order_seed);
-      std::normal_distribution<double> dist(/*mean=*/0, /*stddev=*/1);
+      std::normal_distribution<float> dist(/*mean=*/0, /*stddev=*/1);
 
-      std::vector<std::vector<double>> detector_coords =
+      std::vector<std::vector<float>> detector_coords =
           get_detector_coords(config.dem);
       if (verbose) {
         for (size_t d = 0; d < detector_coords.size(); ++d) {
@@ -238,7 +238,7 @@ struct Args {
           config.det_orders[det_order] = inv_perm;
         }
       } else {
-        std::vector<double> inner_products(config.dem.count_detectors());
+        std::vector<float> inner_products(config.dem.count_detectors());
 
         if (!detector_coords.size() || !detector_coords.at(0).size()) {
           // If there are no detector coordinates, just use the standard ordering
@@ -253,7 +253,7 @@ struct Args {
           // orientation
           for (size_t det_order = 0; det_order < num_det_orders; ++det_order) {
             // Sample a direction
-            std::vector<double> orientation_vector;
+            std::vector<float> orientation_vector;
             for (size_t i = 0; i < detector_coords.at(0).size(); ++i) {
               orientation_vector.push_back(dist(rng));
             }
@@ -574,8 +574,8 @@ int main(int argc, char* argv[]) {
   std::atomic<size_t> next_unclaimed_shot;
   std::vector<std::atomic<bool>> finished(shots.size());
   std::vector<common::ObservablesMask> obs_predicted(shots.size());
-  std::vector<double> cost_predicted(shots.size());
-  std::vector<double> decoding_time_seconds(shots.size());
+  std::vector<float> cost_predicted(shots.size());
+  std::vector<float> decoding_time_seconds(shots.size());
   std::vector<std::atomic<bool>> low_confidence(shots.size());
   std::vector<std::thread> decoder_threads;
   std::vector<std::atomic<size_t>> error_use_totals(config.dem.count_errors());
@@ -627,7 +627,7 @@ int main(int argc, char* argv[]) {
   }
   size_t num_errors = 0;
   size_t num_low_confidence = 0;
-  double total_time_seconds = 0;
+  float total_time_seconds = 0;
   size_t num_observables = config.dem.count_observables();
   size_t shot = 0;
   for (; shot < shots.size(); ++shot) {
